Move a Router users to index
